requirements.txt
affine==2.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==3.0.0
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
executing==2.1.0
fastapi==0.115.5
fonttools==4.55.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.23.0
google-auth==2.36.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.15.1
nest-asyncio==1.6.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.2
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
rasterio==1.4.3
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
streamlit==1.40.2
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
